alignment changes with JIRA 7
This commit bumps the reqwest's version to the latest, needed this to
prevent conflicts with other libraries and bring in the new features it
supports when using client with goji. Changed the api to use 'api' over
'agile' (is agile the old one, do we need to support both). Updated the
fields for issue creation, optional fields are now optional and
custom/additional fields can be added through the use of a BTreeMap.